WIP: Changed model to handle the int64 version of the model weights

+++ src/main/scala/com/johnsnowlabs/ml/ai/util/Generation/Logit/Logit.scala
@@ -18,7 +18,7 @@ package com.johnsnowlabs.ml.ai.util.Generation.Logit
 
 abstract class Logit {
   def call(
-      inputIds: Seq[Array[Int]],
+      inputIds: Seq[Array[Long]],
       scores: Array[Array[Float]],
       currentLength: Int): Array[Array[Float]]
